Make connection aware of the current keyspace (#180)
Closes #179.

Use the keyspace information to add it to "Xandra.Prepared" structs so that it can
be used together with the statement as key in the prepared cache. This avoids
erroneously using queries that were prepared against another keyspace.

To do so, response decoding has been moved inside the connection, akin to what
was already done in "handle_prepare" to populate the prepared cache. The decode
function in query structs is now simply an identity since the result gets there
already decoded.